requirements.txt
asgiref==3.2.10
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
dj-database-url==0.5.0
Django==3.1.2
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djoser==2.0.5
drf-nested-routers==0.92.5
idna==2.10
jwt==1.1.0
pycparser==2.20
pytz==2020.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
v==0.0.0
